`build-logic` refactoring.
- multi-project build
- move convention plugins to `:kstreamlined-gradle-plugin` project
- add new CI job for running `build-logic` checks
