+++ .github/actions/cache-build-logic/action.yml
@@ -8,4 +8,4 @@ runs:
       uses: actions/cache@v4
       with:
         path: build-logic/build
-        key: build-logic-${{ hashFiles('build-logic/src/**', 'build-logic/build.gradle.kts','build-logic/settings.gradle.kts') }}
+        key: build-logic-${{ hashFiles('build-logic/**/src/**', 'build-logic/**/build.gradle.kts','build-logic/settings.gradle.kts') }}
