+++ build-logic/kstreamlined-gradle-plugin/src/main/kotlin/buildEnvironment.kt
