+++ build-logic/kstreamlined-gradle-plugin/src/main/kotlin/productFlavorConfigurationAccessors.kt
