+++ build-logic/kstreamlined-gradle-plugin/src/main/kotlin/variantUtils.kt
